Add in-progress reference parser.
The code in the /reference folder is for a reference implementation of parsing the .TDT file as per this spec.

Consider this highly experimental.

I'm trying to not write an imperative mess, but just attach some metadata and let the code do the rest. This could make later writing easier to implement as well.